ENH: rename connection file
